restructuring and movin around
