+++ modules/AOW_Actions/FormulaCalculator/FormulaCalculator.php
@@ -1,12 +1,11 @@
 <?php
 
 /**
- *
  * SugarCRM Community Edition is a customer relationship management program developed by
  * SugarCRM, Inc. Copyright (C) 2004-2013 SugarCRM Inc.
  *
  * SuiteCRM is an extension to SugarCRM Community Edition developed by SalesAgility Ltd.
- * Copyright (C) 2011 - 2017 SalesAgility Ltd.
+ * Copyright (C) 2011 - 2016 SalesAgility Ltd.
  *
  * This program is free software; you can redistribute it and/or modify it under
  * the terms of the GNU Affero General Public License version 3 as published by the
@@ -38,35 +37,6 @@
  * reasonably feasible for  technical reasons, the Appropriate Legal Notices must
  * display the words  "Powered by SugarCRM" and "Supercharged by SuiteCRM".
  */
-
-/**
- * Class FormulaNode
- */
-class FormulaNode
-{
-    public $text;
-    public $level;
-    public $parent;
-    public $children = array();
-    public $evaluatedValue;
-
-    public function __construct($text, $level, $parent = null)
-    {
-        $this->text = $text;
-        $this->level = $level;
-        $this->parent = $parent;
-    }
-
-    public function addChild($childNode)
-    {
-        $this->children [] = $childNode;
-    }
-
-    public function isLeaf()
-    {
-        return count($this->children) === 0;
-    }
-}
 
 /**
  * Class FormulaCalculator
